src/ai_bootcamp/app/demo/rag/12-RAG/requirements.txt
# RAG Basic PDF 스크립트 실행을 위한 필수 패키지들

# 기본 환경 설정
python-dotenv>=1.0.0

# LangChain 핵심 패키지들
langchain-core>=0.1.0
langchain-openai>=0.1.0
langchain-text-splitters>=0.1.0
langchain-community>=0.1.0

# 벡터 데이터베이스
faiss-cpu>=1.7.0

# 문서 로더 (텍스트 파일용)
# TextLoader는 langchain-community에 포함됨

# OpenAI API (LLM용)
openai>=1.0.0

# 로컬 임베딩 모델 (API 키 없이 사용 가능)
sentence-transformers>=2.2.0
torch>=1.9.0
langchain-huggingface>=0.1.0

# 선택적 패키지들 (LangSmith 추적용)
langchain-teddynote>=0.1.0

# 추가 유틸리티 (선택사항)
numpy>=1.21.0
pandas>=1.3.0 